Add support for multiple integrations
Now, each event integration makes use of its own event
listerer and notifies the associated channel on events.

Fixes #150
